Completed Navbar
Completed nav bar includes: hamburger icon + sidebar and the logo (name) in the top left. The nav bar changes to a solid color after scrolling from the top. The sidebar pulls out when hamburger menu is clicked, and is closed with either the hamburger menu or clicking the main body section.